# docs/swagger.yaml
openapi: 3.0.3
info:
  title: API Invitely
  version: 0.1.0
  description: API para convites de eventos por organização, convidados, confirmação de presença, métricas do painel e análises.
servers:
  - url: http://localhost:8080
    description: Ambiente local
tags:
  - name: Saúde
    description: Verificação do status da API.
  - name: Autenticação
    description: Login, cadastro e consulta do usuário autenticado.
  - name: Eventos
    description: Gestão de eventos da organização autenticada.
  - name: Lembretes
    description: Envio e enfileiramento de lembretes de RSVP por e-mail.
  - name: Orçamento
    description: Itens de orçamento vinculados aos eventos.
  - name: Presentes
    description: Lista de presentes vinculada aos eventos.
  - name: Convidados
    description: Gestão de convidados vinculados aos eventos.
  - name: Painel
    description: Métricas gerais da organização autenticada.
  - name: Análises
    description: Resumos e indicadores por evento.
  - name: Confirmação de Presença
    description: Registro público de confirmação de presença.
  - name: Swagger
    description: Documentação interativa e documentos OpenAPI.
paths:
  /health:
    get:
      tags: [Saúde]
      summary: Verificar saúde da API
      responses:
        "200":
          description: API disponível
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaSaude"
  /auth/login:
    post:
      tags: [Autenticação]
      summary: Entrar na conta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisicaoLogin"
      responses:
        "200":
          description: Sessão autenticada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaAutenticacao"
        "400":
          $ref: "#/components/responses/RequisicaoInvalida"
        "401":
          $ref: "#/components/responses/NaoAutorizado"
  /auth/register:
    post:
      tags: [Autenticação]
      summary: Cadastrar usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisicaoCadastro"
      responses:
        "201":
          description: Usuário cadastrado e sessão criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaAutenticacao"
        "400":
          $ref: "#/components/responses/RequisicaoInvalida"
  /auth/me:
    get:
      tags: [Autenticação]
      summary: Consultar usuário atual
      description: Retorna o usuário associado ao token Bearer enviado no cabeçalho Authorization.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Usuário autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaUsuarioAutenticado"
        "401":
          $ref: "#/components/responses/NaoAutorizado"
  /events:
    get:
      tags: [Eventos]
      summary: Listar eventos
      description: Lista os eventos da organização vinculada ao usuário autenticado.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de eventos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaListaEventos"
        "401":
          $ref: "#/components/responses/NaoAutorizado"
    post:
      tags: [Eventos]
      summary: Criar evento
      description: Cria um evento para a organização vinculada ao usuário autenticado.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisicaoCriarEvento"
      responses:
        "201":
          description: Evento criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaEvento"
        "400":
          $ref: "#/components/responses/RequisicaoInvalida"
        "401":
          $ref: "#/components/responses/NaoAutorizado"
  /events/{id}:
    get:
      tags: [Eventos]
      summary: Buscar evento por ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IDEvento"
      responses:
        "200":
          description: Detalhes do evento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaEvento"
        "401":
          $ref: "#/components/responses/NaoAutorizado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    put:
      tags: [Eventos]
      summary: Atualizar evento
      description: Atualiza os dados principais de um evento da organização autenticada.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IDEvento"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisicaoAtualizarEvento"
      responses:
        "200":
          description: Evento atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaEvento"
        "400":
          $ref: "#/components/responses/RequisicaoInvalida"
        "401":
          $ref: "#/components/responses/NaoAutorizado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    delete:
      tags: [Eventos]
      summary: Excluir evento
      description: Remove um evento da organização autenticada. Convidados, confirmações, orçamento e presentes vinculados são removidos em cascata.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IDEvento"
      responses:
        "204":
          description: Evento excluído
        "401":
          $ref: "#/components/responses/NaoAutorizado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /events/{id}/reminders:
    post:
      tags: [Lembretes]
      summary: Enviar lembretes de RSVP
      description: Valida os dados do lembrete, garante que o evento pertence ao tenant autenticado, registra uma campanha e envia o e-mail pelo Brevo.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IDEvento"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisicaoLembrete"
            example:
              from_email: organizador@example.com
              recipients:
                - convidado1@example.com
                - convidado2@example.com
              subject: "Lembrete: confirme sua presenca no evento"
              message: "Oi! Passando para lembrar voce de confirmar presenca."
      responses:
        "202":
          description: Lembretes enfileirados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaLembrete"
        "400":
          $ref: "#/components/responses/RequisicaoInvalida"
        "401":
          $ref: "#/components/responses/NaoAutorizado"
        "403":
          $ref: "#/components/responses/Proibido"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "422":
          $ref: "#/components/responses/ErroValidacao"
        "502":
          $ref: "#/components/responses/ErroProvedorEmail"
        "503":
          $ref: "#/components/responses/ProvedorEmailIndisponivel"
  /events/{id}/budget:
    get:
      tags: [Orçamento]
      summary: Listar itens de orçamento
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IDEvento"
      responses:
        "200":
          description: Lista de itens de orçamento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaListaItensOrcamento"
        "401":
          $ref: "#/components/responses/NaoAutorizado"
    post:
      tags: [Orçamento]
      summary: Criar item de orçamento
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IDEvento"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisicaoItemOrcamento"
      responses:
        "201":
          description: Item de orçamento criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaItemOrcamento"
        "400":
          $ref: "#/components/responses/RequisicaoInvalida"
        "401":
          $ref: "#/components/responses/NaoAutorizado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /budget/{id}:
    put:
      tags: [Orçamento]
      summary: Atualizar item de orçamento
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IDItemOrcamento"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisicaoItemOrcamento"
      responses:
        "200":
          description: Item de orçamento atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaItemOrcamento"
        "400":
          $ref: "#/components/responses/RequisicaoInvalida"
        "401":
          $ref: "#/components/responses/NaoAutorizado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    delete:
      tags: [Orçamento]
      summary: Excluir item de orçamento
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IDItemOrcamento"
      responses:
        "204":
          description: Item de orçamento excluído
        "401":
          $ref: "#/components/responses/NaoAutorizado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /events/{id}/gifts:
    get:
      tags: [Presentes]
      summary: Listar presentes
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IDEvento"
      responses:
        "200":
          description: Lista de presentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaListaPresentes"
        "401":
          $ref: "#/components/responses/NaoAutorizado"
    post:
      tags: [Presentes]
      summary: Criar presente
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IDEvento"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisicaoPresente"
      responses:
        "201":
          description: Presente criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaPresente"
        "400":
          $ref: "#/components/responses/RequisicaoInvalida"
        "401":
          $ref: "#/components/responses/NaoAutorizado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /gifts/{id}:
    put:
      tags: [Presentes]
      summary: Atualizar presente
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IDPresente"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisicaoPresente"
      responses:
        "200":
          description: Presente atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaPresente"
        "400":
          $ref: "#/components/responses/RequisicaoInvalida"
        "401":
          $ref: "#/components/responses/NaoAutorizado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    delete:
      tags: [Presentes]
      summary: Excluir presente
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IDPresente"
      responses:
        "204":
          description: Presente excluído
        "401":
          $ref: "#/components/responses/NaoAutorizado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /guests:
    get:
      tags: [Convidados]
      summary: Listar convidados por evento
      security:
        - bearerAuth: []
      parameters:
        - name: event_id
          in: query
          required: true
          description: ID do evento.
          schema:
            type: string
          example: evt_123
      responses:
        "200":
          description: Lista de convidados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaListaConvidados"
        "401":
          $ref: "#/components/responses/NaoAutorizado"
    post:
      tags: [Convidados]
      summary: Criar convidado
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisicaoCriarConvidado"
      responses:
        "201":
          description: Convidado criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaConvidado"
        "400":
          $ref: "#/components/responses/RequisicaoInvalida"
        "401":
          $ref: "#/components/responses/NaoAutorizado"
  /dashboard:
    get:
      tags: [Painel]
      summary: Consultar visão geral do painel
      description: Retorna métricas consolidadas da organização vinculada ao usuário autenticado.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Visão geral do painel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaPainel"
        "401":
          $ref: "#/components/responses/NaoAutorizado"
  /analytics/events/{eventID}:
    get:
      tags: [Análises]
      summary: Consultar resumo de confirmações do evento
      security:
        - bearerAuth: []
      parameters:
        - name: eventID
          in: path
          required: true
          description: ID do evento.
          schema:
            type: string
          example: evt_123
      responses:
        "200":
          description: Resumo analítico do evento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaAnaliseEvento"
        "401":
          $ref: "#/components/responses/NaoAutorizado"
  /rsvp:
    post:
      tags: [Confirmação de Presença]
      summary: Enviar confirmação de presença
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisicaoConfirmacaoPresenca"
      responses:
        "200":
          description: Confirmação registrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaConfirmacaoPresenca"
        "400":
          $ref: "#/components/responses/RequisicaoInvalida"
  /swagger:
    get:
      tags: [Swagger]
      summary: Interface Swagger
      responses:
        "200":
          description: Página HTML da documentação Swagger
          content:
            text/html:
              schema:
                type: string
  /swagger/doc.json:
    get:
      tags: [Swagger]
      summary: Documento OpenAPI em JSON
      responses:
        "200":
          description: Documento OpenAPI em JSON
          content:
            application/json:
              schema:
                type: object
  /swagger/doc.yaml:
    get:
      tags: [Swagger]
      summary: Documento OpenAPI em YAML
      responses:
        "200":
          description: Documento OpenAPI em YAML
          content:
            application/yaml:
              schema:
                type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT no formato `Bearer <token>`.
  parameters:
    IDEvento:
      name: id
      in: path
      required: true
      description: ID do evento.
      schema:
        type: string
      example: evt_123
    IDItemOrcamento:
      name: id
      in: path
      required: true
      description: ID do item de orçamento.
      schema:
        type: string
      example: bgt_123
    IDPresente:
      name: id
      in: path
      required: true
      description: ID do presente.
      schema:
        type: string
      example: gft_123
  responses:
    RequisicaoInvalida:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RespostaErro"
    NaoAutorizado:
      description: Token Bearer ausente ou inválido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RespostaErro"
    Proibido:
      description: O evento não pertence à organização autenticada
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RespostaErro"
    NaoEncontrado:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RespostaErro"
    ErroValidacao:
      description: Dados enviados falharam na validação
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RespostaErroValidacao"
    ErroProvedorEmail:
      description: O provedor de e-mail respondeu com erro
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RespostaErro"
    ProvedorEmailIndisponivel:
      description: O provedor de e-mail está indisponível ou não foi configurado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RespostaErro"
  schemas:
    RespostaErro:
      type: object
      properties:
        error:
          type: string
          example: unauthorized
    ErroCampo:
      type: object
      properties:
        field:
          type: string
          example: recipients
        message:
          type: string
          example: recipients must contain at least one email
    RespostaErroValidacao:
      type: object
      properties:
        error:
          type: string
          example: validation failed
        data:
          type: array
          items:
            $ref: "#/components/schemas/ErroCampo"
    RespostaSaude:
      type: object
      properties:
        data:
          type: object
          properties:
            status:
              type: string
              example: ok
    RequisicaoLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: admin@example.com
        password:
          type: string
          format: password
          example: secret123
    RequisicaoCadastro:
      type: object
      required: [email, password, name, role]
      properties:
        email:
          type: string
          format: email
          example: admin@example.com
        password:
          type: string
          format: password
          example: secret123
        name:
          type: string
          example: Usuario Admin
        role:
          type: string
          example: tenant_admin
    UsuarioAutenticado:
      type: object
      properties:
        id:
          type: string
        tenant_id:
          type: string
        supabase_user_id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
    DadosAutenticacao:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          example: 3600
        user:
          $ref: "#/components/schemas/UsuarioAutenticado"
    RespostaAutenticacao:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/DadosAutenticacao"
    RespostaUsuarioAutenticado:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/UsuarioAutenticado"
    RequisicaoCriarEvento:
      type: object
      required: [title, starts_at, ends_at]
      properties:
        title:
          type: string
          example: Lancamento do Produto
        description:
          type: string
          example: Evento exclusivo para convidados.
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        location:
          type: string
          example: Sao Paulo
    RequisicaoAtualizarEvento:
      type: object
      required: [title, starts_at, ends_at]
      properties:
        title:
          type: string
          example: Lancamento do Produto Atualizado
        description:
          type: string
          example: Evento atualizado para convidados.
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        location:
          type: string
          example: Rio de Janeiro
    Evento:
      type: object
      properties:
        id:
          type: string
        tenant_id:
          type: string
        title:
          type: string
        description:
          type: string
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        location:
          type: string
        slug:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RespostaEvento:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Evento"
    RespostaListaEventos:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Evento"
    RequisicaoLembrete:
      type: object
      required: [from_email, recipients, subject, message]
      properties:
        from_email:
          type: string
          format: email
          example: organizador@example.com
        recipients:
          type: array
          minItems: 1
          maxItems: 200
          items:
            type: string
            format: email
          example:
            - convidado1@example.com
            - convidado2@example.com
        subject:
          type: string
          example: "Lembrete: confirme sua presenca no evento"
        message:
          type: string
          example: "Oi! Passando para lembrar voce de confirmar presenca."
    LembreteEnfileirado:
      type: object
      properties:
        campaign_id:
          type: string
          example: cmp_123
        event_id:
          type: string
          example: evt_123
        queued:
          type: integer
          example: 2
        status:
          type: string
          example: sent
    RespostaLembrete:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/LembreteEnfileirado"
    RequisicaoItemOrcamento:
      type: object
      required: [description, amount]
      properties:
        description:
          type: string
          example: Buffet
        category:
          type: string
          example: Alimentacao
        amount:
          type: number
          format: double
          minimum: 0
          example: 2500
        paid:
          type: boolean
          example: false
    ItemOrcamento:
      type: object
      properties:
        id:
          type: string
        event_id:
          type: string
        description:
          type: string
        category:
          type: string
        amount:
          type: number
          format: double
        paid:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RespostaItemOrcamento:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/ItemOrcamento"
    RespostaListaItensOrcamento:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ItemOrcamento"
    RequisicaoPresente:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Jogo de panelas
        description:
          type: string
          example: Presente sugerido para a casa nova.
        price:
          type: number
          format: double
          minimum: 0
          example: 399.9
        url:
          type: string
          example: https://loja.example.com/produto
        reserved:
          type: boolean
          example: false
        reserved_by:
          type: string
          example: Maria Silva
    Presente:
      type: object
      properties:
        id:
          type: string
        event_id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        url:
          type: string
        reserved:
          type: boolean
        reserved_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RespostaPresente:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Presente"
    RespostaListaPresentes:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Presente"
    RequisicaoCriarConvidado:
      type: object
      required: [event_id, name]
      properties:
        event_id:
          type: string
          example: evt_123
        name:
          type: string
          example: Maria Silva
        email:
          type: string
          format: email
          example: maria@example.com
        phone:
          type: string
          example: "+5511999999999"
    Convidado:
      type: object
      properties:
        id:
          type: string
        event_id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        status:
          type: string
          example: pending
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RespostaConvidado:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Convidado"
    RespostaListaConvidados:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Convidado"
    VisaoGeralPainel:
      type: object
      properties:
        total_events:
          type: integer
        total_guests:
          type: integer
        total_confirmed:
          type: integer
    RespostaPainel:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/VisaoGeralPainel"
    ResumoEvento:
      type: object
      properties:
        event_id:
          type: string
        guest_count:
          type: integer
        confirmed_count:
          type: integer
        declined_count:
          type: integer
        pending_count:
          type: integer
    RespostaAnaliseEvento:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/ResumoEvento"
    RequisicaoConfirmacaoPresenca:
      type: object
      required: [guest_id, event_id, status]
      properties:
        guest_id:
          type: string
          example: gst_123
        event_id:
          type: string
          example: evt_123
        status:
          type: string
          enum: [confirmed, declined, pending]
    ConfirmacaoPresenca:
      type: object
      properties:
        id:
          type: string
        guest_id:
          type: string
        event_id:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RespostaConfirmacaoPresenca:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/ConfirmacaoPresenca"